Remove requirement for global gulp and rollup
Requiring global installation of any tools is not the best practice, as it can
introduce situations when one project requires one version of e.g. gulp and
another application requires another version. To combat that, npm provides a great
way to encapsulate dependencies inside `node_modules` directory and includes all
binaries in PATH whenever `npm ...` commands are ran.

In this commit, `rollup` was added as a dependency to `package.json` so that global
installation is not required. Additionally, npm scripts for common operations were
added: `npm start` for development and `npm run build` for one-time building.

README was updated to match the updates.

Took 5 minutes